[SEARCH-16] File update / removal

+++ cern_search_rest_api/modules/cernsearch/files.py
@@ -14,7 +14,7 @@ from cern_search_rest_api.modules.cernsearch.api import CernSearchRecord
 from cern_search_rest_api.modules.cernsearch.errors import ObjectNotFoundError
 from flask import current_app
 from invenio_db import db
-from invenio_files_rest.models import FileInstance, ObjectVersion
+from invenio_files_rest.models import Bucket, FileInstance, ObjectVersion
 from invenio_records_files.api import FilesIterator
 from invenio_records_files.models import RecordsBuckets
 
@@ -54,41 +54,76 @@ def delete_previous_record_file_if_exists(obj: ObjectVersion):
 
     __delete_all_files_except(record.files, obj)
     __delete_all_files_except(record.files_content, obj)
-    __delete_object_versions_except(obj)
 
 
-def __delete_object_versions_except(obj: ObjectVersion):
-    for version in ObjectVersion.get_versions(obj.bucket, obj.key):
-        if version.version_id != obj.version_id:
-            delete_file(version)
+def delete_object_version(obj: ObjectVersion):
+    """Delete file on filesystem and soft delete on database."""
+    current_app.logger.debug(f"Delete Object Version: {str(obj)}")
+
+    #  Soft delete bucket
+    obj.delete(obj.bucket, obj.key)
+
+    delete_file_instance(obj)
+
+    db.session.commit()
 
 
-def __delete_all_files_except(objects: FilesIterator, obj: ObjectVersion):
-    for file in objects:
-        if file.obj.version_id != obj.version_id:
-            delete_file(file.obj)
-
-
-def delete_file(obj: ObjectVersion):
-    """Delete file on filesystem and database."""
-    current_app.logger.debug(f"Delete File: {str(obj)}")
-
-    obj.remove()
+def delete_file_instance(obj: ObjectVersion):
+    """Delete file on filesystem and mark as not readable."""
+    current_app.logger.debug(f"Delete file instance: {str(obj)}")
 
     if obj.file_id:
-        # First remove FileInstance from database and commit transaction to
-        # ensure integrity constraints are checked and enforced.
-        f = FileInstance.get(str(obj.file_id))
+        f = FileInstance.get(str(obj.file_id))  # type: FileInstance
 
         if not f:
             raise ObjectNotFoundError(f"FileInstance with id {obj.file_id}")
 
-        f.delete()
-        # Next, remove the file on disk. This leaves the possibility of having
-        # a file on disk dangling in case the database removal works, and the
-        # disk file removal doesn't work.
-        f.storage().delete()
+        is_readable = f.readable
+
+        # Mark file not readable
+        f.readable = False
+
+        # Remove the file on disk
+        if is_readable:
+            f.storage().delete()
 
     db.session.commit()
 
-    # file_deleted.send(obj)
+
+def delete_record_file(obj: ObjectVersion):
+    """Delete associated file to record."""
+    record = record_from_object_version(obj)  # type: CernSearchRecord
+    current_app.logger.debug(f"Cleanup file: {str(obj)}")
+
+    delete_object_version(obj)
+    if obj.key in record.files_content:
+        delete_object_version(record.files_content[obj.key])
+
+
+def delete_all_record_files(record: CernSearchRecord):
+    """Delete all associated files to record."""
+    current_app.logger.debug(f"Cleanup files: {str(record)}")
+
+    __delete_all_files(record.files)
+    __delete_all_files(record.files_content)
+
+
+def __delete_all_files(objects: FilesIterator):
+    for file in objects:
+        delete_object_version(file.obj)
+
+
+def __delete_all_files_except(objects: FilesIterator, obj: ObjectVersion):
+    for file in objects:
+        if file.obj.key == obj.key:
+            __delete_object_versions_except(obj, objects.bucket)
+
+            continue
+
+        delete_object_version(file.obj)
+
+
+def __delete_object_versions_except(obj: ObjectVersion, bucket: Bucket):
+    for version in ObjectVersion.get_versions(bucket, obj.key):
+        if version.version_id != obj.version_id:
+            delete_file_instance(version)
